backend/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
python-multipart
Pillow
moviepy
imageio
